Changing a direction from "left" to "right"
Under ".grid-container" we have the following:
"For our container, we are obviously using `display: grid` to render the container as a CSS Grid. But the next property might be unfamiliar to you: `resize: both`. This is a property that allows the user to resize the container by clicking and dragging from the bottom LEFT corner."

I changed the word "LEFT" to "RIGHT" since I believe that is what it's supposed to say.